MAJOR: Disable ability, damage-class, egg-group, item, move, type, og, splash routes. Disable Shell
