+++ .github/workflows/nextjs.yml
@@ -56,8 +56,8 @@ jobs:
             ${{ runner.os }}-nextjs-${{ hashFiles('**/package-lock.json', '**/pnpm-lock.yaml', '**/yarn.lock', '**/bun.lock') }}-
       - name: Install dependencies
         run: bun install
-      - name: Install Playwright browsers
-        run: bunx playwright install
+      # - name: Install Playwright browsers
+      #   run: bunx playwright install
       - name: Build with Next.js
         run: bun next build
       - name: Upload artifact
