+++ app/(main)/_damage-class/[slug]/page.tsx
