+++ app/(main)/_type/[slug]/page.tsx
