+++ app/(main)/layout.tsx
@@ -1,6 +1,5 @@
-import { Suspense } from 'react'
+import Header from '@/components/shared/Header'
 import Footer from '@/components/shared/Footer'
-import Shell from '@/components/shared/Shell'
 import { VersionGroupProvider } from '@/components/shared/VersionGroupProvider'
 
 export default function RootLayout({
@@ -11,9 +10,10 @@ export default function RootLayout({
   return (
     <VersionGroupProvider>
       <div className="flex min-h-screen flex-col gap-6">
-        <Suspense>
+        {/* <Suspense>
           <Shell />
-        </Suspense>
+        </Suspense> */}
+        <Header />
         <main className="flex grow flex-col items-center">{children}</main>
         <Footer />
       </div>
