+++ app/_(splash)/splash/[slug]/[[...variant]]/page.tsx
