+++ components/shared/Shell.tsx
@@ -22,129 +22,129 @@ import {
 } from '@/lib/utils/pokeapiHelpers'
 
 export default async function Shell() {
-  const speciesList =
-    process?.env?.NODE_ENV && process?.env?.NODE_ENV === 'development'
-      ? await getTestSpeciesList()
-      : await pokeapi.getPokemonSpeciesList({
-          limit: 151,
-          offset: 0,
-        })
+  // const speciesList =
+  //   process?.env?.NODE_ENV && process?.env?.NODE_ENV === 'development'
+  //     ? await getTestSpeciesList()
+  //     : await pokeapi.getPokemonSpeciesList({
+  //         limit: 151,
+  //         offset: 0,
+  //       })
 
-  const movesList = await getTestMovesList()
-  const abilitiesList = await getTestAbilitiesList()
-  const itemsList = await getTestItemsList()
-  const eggGroupsList = await pokeapi.getEggGroupsList()
+  // const movesList = await getTestMovesList()
+  // const abilitiesList = await getTestAbilitiesList()
+  // const itemsList = await getTestItemsList()
+  // const eggGroupsList = await pokeapi.getEggGroupsList()
 
-  const species = (await batchFetch(
-    speciesList.results.map((result) => result.url),
-    (url) => pokeapi.getResource(url),
-    10
-  )) as PokemonSpecies[]
-  // const species = await pokeapi.getPokemonSpeciesByName(
-  //   speciesList.results.map((resource: NamedAPIResource) => resource.name)
+  // const species = (await batchFetch(
+  //   speciesList.results.map((result) => result.url),
+  //   (url) => pokeapi.getResource(url),
+  //   10
+  // )) as PokemonSpecies[]
+  // // const species = await pokeapi.getPokemonSpeciesByName(
+  // //   speciesList.results.map((resource: NamedAPIResource) => resource.name)
+  // // )
+  // const monsters = (await batchFetch(species, getMonstersBySpecies, 10)).flat()
+  // const moves = await pokeapi.getMoveByName(
+  //   movesList.results.map((resource: NamedAPIResource) => resource.name)
   // )
-  const monsters = (await batchFetch(species, getMonstersBySpecies, 10)).flat()
-  const moves = await pokeapi.getMoveByName(
-    movesList.results.map((resource: NamedAPIResource) => resource.name)
-  )
-  const abilities = await pokeapi.getAbilityByName(
-    abilitiesList.results.map((resource: NamedAPIResource) => resource.name)
-  )
-  const items = await pokeapi.getItemByName(
-    itemsList.results.map((resource: NamedAPIResource) => resource.name)
-  )
-  const types = await pokeapi.getTypeByName(Object.values(TypeKey))
-  const damageClass = await pokeapi.getMoveDamageClassByName(
-    Object.values(DamageClassKey)
-  )
-  const eggGroups = await pokeapi.getEggGroupByName(
-    eggGroupsList.results.map((resource: NamedAPIResource) => resource.name)
-  )
+  // const abilities = await pokeapi.getAbilityByName(
+  //   abilitiesList.results.map((resource: NamedAPIResource) => resource.name)
+  // )
+  // const items = await pokeapi.getItemByName(
+  //   itemsList.results.map((resource: NamedAPIResource) => resource.name)
+  // )
+  // const types = await pokeapi.getTypeByName(Object.values(TypeKey))
+  // const damageClass = await pokeapi.getMoveDamageClassByName(
+  //   Object.values(DamageClassKey)
+  // )
+  // const eggGroups = await pokeapi.getEggGroupByName(
+  //   eggGroupsList.results.map((resource: NamedAPIResource) => resource.name)
+  // )
 
-  const monsterItems: IndexItem[] = await Promise.all(
-    monsters.map(async (monster) => {
-      const { id, key, name, species, pokemon, form } = monster
-      const imageId = id.toString().padStart(4, '0')
-      const imageUrl =
-        pokemon.sprites.other?.home?.front_default ??
-        pokemon.sprites.other?.['official-artwork'].front_default ??
-        `https://resource.pokemon-home.com/battledata/img/pokei128/icon${imageId}_f00_s0.png`
-      const title = `${name} #${imageId}`
-      const path = form ? `/${species.name}/${form.name}` : `/${species.name}`
+  // const monsterItems: IndexItem[] = await Promise.all(
+  //   monsters.map(async (monster) => {
+  //     const { id, key, name, species, pokemon, form } = monster
+  //     const imageId = id.toString().padStart(4, '0')
+  //     const imageUrl =
+  //       pokemon.sprites.other?.home?.front_default ??
+  //       pokemon.sprites.other?.['official-artwork'].front_default ??
+  //       `https://resource.pokemon-home.com/battledata/img/pokei128/icon${imageId}_f00_s0.png`
+  //     const title = `${name} #${imageId}`
+  //     const path = form ? `/${species.name}/${form.name}` : `/${species.name}`
 
-      return {
-        id,
-        title,
-        slug: key,
-        path,
-        // imageUrl,
-      } as IndexItem
-    })
-  )
+  //     return {
+  //       id,
+  //       title,
+  //       slug: key,
+  //       path,
+  //       // imageUrl,
+  //     } as IndexItem
+  //   })
+  // )
 
-  const moveItems: IndexItem[] = moves.map((move) => ({
-    id: move.id,
-    title: getTranslation(move.names, 'name')!,
-    slug: move.name,
-    path: `/move/${move.name}`,
-  }))
+  // const moveItems: IndexItem[] = moves.map((move) => ({
+  //   id: move.id,
+  //   title: getTranslation(move.names, 'name')!,
+  //   slug: move.name,
+  //   path: `/move/${move.name}`,
+  // }))
 
-  const abilityItems: IndexItem[] = abilities.map((ability) => ({
-    id: ability.id,
-    title: getTranslation(ability.names, 'name')!,
-    slug: ability.name,
-    path: `/ability/${ability.name}`,
-  }))
+  // const abilityItems: IndexItem[] = abilities.map((ability) => ({
+  //   id: ability.id,
+  //   title: getTranslation(ability.names, 'name')!,
+  //   slug: ability.name,
+  //   path: `/ability/${ability.name}`,
+  // }))
 
-  const itemItems: IndexItem[] = items.map((item) => ({
-    id: item.id,
-    title: getTranslation(item.names, 'name')!,
-    slug: item.name,
-    path: `/item/${item.name}`,
-  }))
+  // const itemItems: IndexItem[] = items.map((item) => ({
+  //   id: item.id,
+  //   title: getTranslation(item.names, 'name')!,
+  //   slug: item.name,
+  //   path: `/item/${item.name}`,
+  // }))
 
-  const typeItems: IndexItem[] = types.map((type) => ({
-    id: type.id,
-    title: TypeLabels[type.name as TypeKey],
-    slug: type.name,
-    path: `/type/${type.name}`,
-  }))
+  // const typeItems: IndexItem[] = types.map((type) => ({
+  //   id: type.id,
+  //   title: TypeLabels[type.name as TypeKey],
+  //   slug: type.name,
+  //   path: `/type/${type.name}`,
+  // }))
 
-  const damageClassItems: IndexItem[] = damageClass.map((damageClass) => {
-    return {
-      id: damageClass.id,
-      title: DamageClassLabels[damageClass.name as DamageClassKey],
-      slug: damageClass.name,
-      path: `/damage-class/${damageClass.name}`,
-    } as IndexItem
-  })
+  // const damageClassItems: IndexItem[] = damageClass.map((damageClass) => {
+  //   return {
+  //     id: damageClass.id,
+  //     title: DamageClassLabels[damageClass.name as DamageClassKey],
+  //     slug: damageClass.name,
+  //     path: `/damage-class/${damageClass.name}`,
+  //   } as IndexItem
+  // })
 
-  const eggGroupItems: IndexItem[] = eggGroups.map((eggGroup) => {
-    const title = getTranslation(eggGroup.names, 'name')!
-    const imageUrl =
-      'https://resource.pokemon-home.com/battledata/img/pokei128/icon0000_f00_s0.png'
-    return {
-      id: eggGroup.id,
-      title,
-      slug: eggGroup.name,
-      path: `/egg-group/${eggGroup.name}`,
-      // imageUrl,
-    } as IndexItem
-  })
+  // const eggGroupItems: IndexItem[] = eggGroups.map((eggGroup) => {
+  //   const title = getTranslation(eggGroup.names, 'name')!
+  //   const imageUrl =
+  //     'https://resource.pokemon-home.com/battledata/img/pokei128/icon0000_f00_s0.png'
+  //   return {
+  //     id: eggGroup.id,
+  //     title,
+  //     slug: eggGroup.name,
+  //     path: `/egg-group/${eggGroup.name}`,
+  //     // imageUrl,
+  //   } as IndexItem
+  // })
 
-  const allItems = [
-    ...monsterItems,
-    ...moveItems,
-    ...abilityItems,
-    ...itemItems,
-    ...typeItems,
-    ...damageClassItems,
-    ...eggGroupItems,
-  ]
+  // const allItems = [
+  //   ...monsterItems,
+  //   ...moveItems,
+  //   ...abilityItems,
+  //   ...itemItems,
+  //   ...typeItems,
+  //   ...damageClassItems,
+  //   ...eggGroupItems,
+  // ]
 
   return (
-    <GlobalIndexProvider indexItems={allItems}>
+    // <GlobalIndexProvider indexItems={allItems}>
       <Header />
-    </GlobalIndexProvider>
+    // </GlobalIndexProvider>
   )
 }
