+++ package.json
@@ -13,6 +13,7 @@
     "@radix-ui/react-dialog": "^1.1.6",
     "@radix-ui/react-visually-hidden": "^1.1.2",
     "@tanstack/react-table": "^8.21.2",
+    "@vercel/analytics": "^1.5.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.4",
     "cva": "^1.0.0-beta.3",
@@ -21,7 +22,6 @@
     "motion": "^12.4.10",
     "next": "^15.2.1",
     "next-themes": "^0.4.4",
-    "playwright": "^1.50.1",
     "pokedex-promise-v2": "^4.2.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
